refactor: remove config store shim, import directly from panoptes.utils.config.store
All 39 files that imported from panoptes.pocs.config.store now import
directly from panoptes.utils.config.store. The intermediate shim
(src/panoptes/pocs/config/store.py) has been deleted — there was no
public release to maintain backward-compatibility with.
