+++ src/panoptes/pocs/camera/simulator/ccd.py
@@ -15,7 +15,7 @@ import astropy.units as u
 
 from panoptes.pocs.camera.sdk import AbstractSDKCamera, AbstractSDKDriver
 from panoptes.pocs.camera.simulator.dslr import Camera as SimCamera
-from panoptes.pocs.config.store import get_config
+from panoptes.utils.config.store import get_config
 
 
 class SDKDriver(AbstractSDKDriver):
