+++ src/panoptes/pocs/dome/__init__.py
@@ -3,7 +3,7 @@ from abc import ABCMeta, abstractmethod
 from panoptes.utils.library import load_module
 
 from panoptes.pocs.base import PanBase
-from panoptes.pocs.config.store import get_config
+from panoptes.utils.config.store import get_config
 from panoptes.pocs.utils.logger import get_logger
 
 logger = get_logger()
