+++ tests/scheduler/test_dispatch_scheduler.py
@@ -9,7 +9,7 @@ from astropy import units as u
 from astropy.coordinates import EarthLocation
 from astropy.time import Time
 
-from panoptes.pocs.config.store import get_config
+from panoptes.utils.config.store import get_config
 from panoptes.pocs.scheduler.constraint import Duration, MoonAvoidance
 from panoptes.pocs.scheduler.dispatch import Scheduler
