+++ tests/test_mount_simulator.py
@@ -7,7 +7,7 @@ from astropy.coordinates import EarthLocation, SkyCoord
 from panoptes.utils import error
 from panoptes.utils.utils import altaz_to_radec
 
-from panoptes.pocs.config.store import get_config
+from panoptes.utils.config.store import get_config
 from panoptes.pocs.mount.simulator import Mount
